ac/surface: use uint16_t for mipmap level pitches

+++ src/amd/common/ac_surface.h
@@ -155,7 +155,7 @@ struct gfx9_surf_layout {
     /* Mipmap level offset within the slice in bytes. Only valid for LINEAR. */
     uint32_t                    offset[RADEON_SURF_MAX_LEVELS];
     /* Mipmap level pitch in elements. Only valid for LINEAR. */
-    uint32_t                    pitch[RADEON_SURF_MAX_LEVELS];
+    uint16_t                    pitch[RADEON_SURF_MAX_LEVELS];
 
     uint64_t                    stencil_offset; /* separate stencil */
